requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
async-timeout==5.0.1
cffi==2.0.0
click==8.2.1
contextlib2==21.6.0
cryptography==45.0.7
diff-match-patch==20241021
Django==5.0
django-cors-headers==4.8.0
django-import-export==4.3.9
django-maintenance-mode==0.22.0
django-ninja==1.4.3
django-ninja-extra==0.30.1
django-ninja-jwt==5.3.7
django-ninja-passwordreset==1.2.5
django-redis==6.0.0
django-simple-history==3.10.1
django-unfold==0.65.0
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
h11==0.16.0
idna==3.10
injector==0.22.0
ninja-schema==0.14.3
openpyxl==3.1.5
psycopg2==2.9.10
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
python-decouple==3.8
python-fsutil==0.15.0
redis==6.4.0
sorl-thumbnail==12.11.0
sqlparse==0.5.3
tablib==3.8.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
